+++ src/main/java/com/xzavier0722/mc/plugin/slimefun4/storage/adapter/mysql/MysqlAdapter.java
@@ -99,241 +99,4 @@ public class MysqlAdapter extends SqlCommonAdapter<MysqlConfig> {
     public void deleteData(RecordKey key) {
         executeSql("DELETE FROM " + mapTable(key.getScope()) + SqlUtils.buildConditionStr(key.getConditions()) + ";");
     }
-
-    private void createProfileTables() {
-        createProfileTable();
-        createResearchTable();
-        createBackpackTable();
-        createBackpackInventoryTable();
-    }
-
-    private void createBlockStorageTables() {
-        createBlockRecordTable();
-        createBlockDataTable();
-        createBlockInvTable();
-        createChunkDataTable();
-        createUniversalInventoryTable();
-        createUniversalRecordTable();
-    }
-
-    private void createProfileTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + profileTable
-                + "("
-                + FIELD_PLAYER_UUID
-                + " CHAR(64) PRIMARY KEY NOT NULL, "
-                + FIELD_PLAYER_NAME
-                + " CHAR(64) NOT NULL, "
-                + FIELD_BACKPACK_NUM
-                + " INT UNSIGNED DEFAULT 0, "
-                + "INDEX index_player_name ("
-                + FIELD_PLAYER_NAME
-                + ")"
-                + ");");
-    }
-
-    private void createResearchTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + researchTable
-                + "("
-                + FIELD_PLAYER_UUID
-                + " CHAR(64) NOT NULL, "
-                + FIELD_RESEARCH_KEY
-                + " CHAR(64) NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "REFERENCES "
-                + profileTable
-                + "("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "INDEX index_player_research ("
-                + FIELD_PLAYER_UUID
-                + ", "
-                + FIELD_RESEARCH_KEY
-                + ")"
-                + ");");
-    }
-
-    private void createBackpackTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + backpackTable
-                + "("
-                + FIELD_BACKPACK_ID
-                + " CHAR(64) PRIMARY KEY NOT NULL, "
-                + FIELD_PLAYER_UUID
-                + " CHAR(64) NOT NULL, "
-                + FIELD_BACKPACK_NUM
-                + " INT UNSIGNED NOT NULL, "
-                + FIELD_BACKPACK_NAME
-                + " CHAR(64) NULL, "
-                + FIELD_BACKPACK_SIZE
-                + " TINYINT UNSIGNED NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "REFERENCES "
-                + profileTable
-                + "("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "INDEX index_player_backpack ("
-                + FIELD_PLAYER_UUID
-                + ", "
-                + FIELD_BACKPACK_NUM
-                + ")"
-                + ");");
-    }
-
-    private void createBackpackInventoryTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + bpInvTable
-                + "("
-                + FIELD_BACKPACK_ID
-                + " CHAR(64) NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " TINYINT UNSIGNED NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_BACKPACK_ID
-                + ") "
-                + "REFERENCES "
-                + backpackTable
-                + "("
-                + FIELD_BACKPACK_ID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_BACKPACK_ID
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createBlockRecordTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + blockRecordTable
-                + "("
-                + FIELD_LOCATION
-                + " CHAR(64) PRIMARY KEY NOT NULL, "
-                + FIELD_CHUNK
-                + " CHAR(64) NOT NULL, "
-                + FIELD_SLIMEFUN_ID
-                + " CHAR(64) NOT NULL, "
-                + "INDEX index_ticking ("
-                + FIELD_CHUNK
-                + ")"
-                + ");");
-    }
-
-    private void createBlockDataTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + blockDataTable
-                + "("
-                + FIELD_LOCATION
-                + " CHAR(64) NOT NULL, "
-                + FIELD_DATA_KEY
-                + " CHAR(64) NOT NULL, "
-                + FIELD_DATA_VALUE
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_LOCATION
-                + ") "
-                + "REFERENCES "
-                + blockRecordTable
-                + "("
-                + FIELD_LOCATION
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_LOCATION
-                + ", "
-                + FIELD_DATA_KEY
-                + ")"
-                + ");");
-    }
-
-    private void createChunkDataTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + chunkDataTable
-                + "("
-                + FIELD_CHUNK
-                + " CHAR(64) NOT NULL, "
-                + FIELD_DATA_KEY
-                + " CHAR(64) NOT NULL, "
-                + FIELD_DATA_VALUE
-                + " TEXT NOT NULL, "
-                + "PRIMARY KEY ("
-                + FIELD_CHUNK
-                + ", "
-                + FIELD_DATA_KEY
-                + ")"
-                + ");");
-    }
-
-    private void createUniversalInventoryTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + universalInvTable
-                + "("
-                + FIELD_UNIVERSAL_UUID
-                + " CHAR(64) NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " TINYINT UNSIGNED NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL,"
-                + "PRIMARY KEY ("
-                + FIELD_UNIVERSAL_UUID
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createBlockInvTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + blockInvTable
-                + "("
-                + FIELD_LOCATION
-                + " CHAR(64) NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " TINYINT UNSIGNED NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_LOCATION
-                + ") "
-                + "REFERENCES "
-                + blockRecordTable
-                + "("
-                + FIELD_LOCATION
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_LOCATION
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createUniversalRecordTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + universalRecordTable
-                + "("
-                + FIELD_UNIVERSAL_UUID
-                + " UUID NOT NULL, "
-                + FIELD_SLIMEFUN_ID
-                + " TEXT NOT NULL, "
-                + FIELD_LAST_PRESENT
-                + " TEXT NOT NULL, "
-                + "PRIMARY KEY ("
-                + FIELD_UNIVERSAL_UUID
-                + ")"
-                + ");");
-    }
 }
