+++ src/main/java/com/xzavier0722/mc/plugin/slimefun4/storage/adapter/postgresql/PostgreSqlAdapter.java
@@ -1,22 +1,5 @@
 package com.xzavier0722.mc.plugin.slimefun4.storage.adapter.postgresql;
 
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_BACKPACK_ID;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_BACKPACK_NAME;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_BACKPACK_NUM;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_BACKPACK_SIZE;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_CHUNK;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_DATA_KEY;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_DATA_VALUE;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_INVENTORY_ITEM;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_INVENTORY_SLOT;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_LAST_PRESENT;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_LOCATION;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_PLAYER_NAME;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_PLAYER_UUID;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_RESEARCH_KEY;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_SLIMEFUN_ID;
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_UNIVERSAL_UUID;
-
 import com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlCommonAdapter;
 import com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlUtils;
 import com.xzavier0722.mc.plugin.slimefun4.storage.common.DataScope;
@@ -125,273 +108,5 @@ public class PostgreSqlAdapter extends SqlCommonAdapter<PostgreSqlConfig> {
     @Override
     public void deleteData(RecordKey key) {
         executeSql("DELETE FROM " + mapTable(key.getScope()) + SqlUtils.buildConditionStr(key.getConditions()) + ";");
-    }
-
-    private void createProfileTables() {
-        createProfileTable();
-        createResearchTable();
-        createBackpackTable();
-        createBackpackInventoryTable();
-    }
-
-    private void createBlockStorageTables() {
-        createBlockRecordTable();
-        createBlockDataTable();
-        createBlockInvTable();
-        createChunkDataTable();
-        createUniversalInventoryTable();
-        createUniversalRecordTable();
-    }
-
-    private void createProfileTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + profileTable
-                + "("
-                + FIELD_PLAYER_UUID
-                + " VARCHAR(64) PRIMARY KEY NOT NULL, "
-                + FIELD_PLAYER_NAME
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_BACKPACK_NUM
-                + " INT DEFAULT 0"
-                + ");");
-
-        executeSql("CREATE INDEX IF NOT EXISTS index_player_name ON " + profileTable + " (" + FIELD_PLAYER_NAME + ");");
-    }
-
-    private void createResearchTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + researchTable
-                + "("
-                + FIELD_PLAYER_UUID
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_RESEARCH_KEY
-                + " VARCHAR(64) NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "REFERENCES "
-                + profileTable
-                + "("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE"
-                + ");");
-
-        executeSql("CREATE INDEX IF NOT EXISTS index_player_research ON "
-                + researchTable
-                + " ("
-                + FIELD_PLAYER_UUID
-                + ", "
-                + FIELD_RESEARCH_KEY
-                + ");");
-    }
-
-    private void createBackpackTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + backpackTable
-                + "("
-                + FIELD_BACKPACK_ID
-                + " VARCHAR(64) PRIMARY KEY NOT NULL, "
-                + FIELD_PLAYER_UUID
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_BACKPACK_NUM
-                + " INT NOT NULL, "
-                + FIELD_BACKPACK_NAME
-                + " VARCHAR(64) NULL, "
-                + FIELD_BACKPACK_SIZE
-                + " SMALLINT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "REFERENCES "
-                + profileTable
-                + "("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE"
-                + ");");
-
-        executeSql("CREATE INDEX IF NOT EXISTS index_player_backpack ON "
-                + backpackTable
-                + " ("
-                + FIELD_PLAYER_UUID
-                + ", "
-                + FIELD_BACKPACK_NUM
-                + ");");
-    }
-
-    private void createBackpackInventoryTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + bpInvTable
-                + "("
-                + FIELD_BACKPACK_ID
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " SMALLINT NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_BACKPACK_ID
-                + ") "
-                + "REFERENCES "
-                + backpackTable
-                + "("
-                + FIELD_BACKPACK_ID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_BACKPACK_ID
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createBlockRecordTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + blockRecordTable
-                + "("
-                + FIELD_LOCATION
-                + " VARCHAR(64) PRIMARY KEY NOT NULL, "
-                + FIELD_CHUNK
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_SLIMEFUN_ID
-                + " VARCHAR(64) NOT NULL"
-                + ");");
-
-        executeSql("CREATE INDEX IF NOT EXISTS index_ticking ON " + blockRecordTable + " (" + FIELD_CHUNK + ");");
-    }
-
-    private void createBlockDataTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + blockDataTable
-                + "("
-                + FIELD_LOCATION
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_DATA_KEY
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_DATA_VALUE
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_LOCATION
-                + ") "
-                + "REFERENCES "
-                + blockRecordTable
-                + "("
-                + FIELD_LOCATION
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_LOCATION
-                + ", "
-                + FIELD_DATA_KEY
-                + ")"
-                + ");");
-    }
-
-    private void createChunkDataTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + chunkDataTable
-                + "("
-                + FIELD_CHUNK
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_DATA_KEY
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_DATA_VALUE
-                + " TEXT NOT NULL, "
-                + "PRIMARY KEY ("
-                + FIELD_CHUNK
-                + ", "
-                + FIELD_DATA_KEY
-                + ")"
-                + ");");
-    }
-
-    private void createBlockInvTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + blockInvTable
-                + "("
-                + FIELD_LOCATION
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " SMALLINT NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_LOCATION
-                + ") "
-                + "REFERENCES "
-                + blockRecordTable
-                + "("
-                + FIELD_LOCATION
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_LOCATION
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createUniversalInventoryTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + universalInvTable
-                + "("
-                + FIELD_UNIVERSAL_UUID
-                + " UUID NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " TINYINT UNSIGNED NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL,"
-                + "PRIMARY KEY ("
-                + FIELD_UNIVERSAL_UUID
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createUniversalRecordTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + universalRecordTable
-                + "("
-                + FIELD_UNIVERSAL_UUID
-                + " UUID NOT NULL, "
-                + FIELD_SLIMEFUN_ID
-                + " TEXT NOT NULL, "
-                + FIELD_LAST_PRESENT
-                + " TEXT NOT NULL, "
-                + "PRIMARY KEY ("
-                + FIELD_UNIVERSAL_UUID
-                + ")"
-                + ");");
-    }
-
-    private void createUniversalDataTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + universalDataTable
-                + "("
-                + FIELD_UNIVERSAL_UUID
-                + " UUID NOT NULL, "
-                + FIELD_DATA_KEY
-                + " VARCHAR(64) NOT NULL, "
-                + FIELD_DATA_VALUE
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_UNIVERSAL_UUID
-                + ") "
-                + "REFERENCES "
-                + universalRecordTable
-                + "("
-                + FIELD_UNIVERSAL_UUID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_UNIVERSAL_UUID
-                + ", "
-                + FIELD_DATA_KEY
-                + ")"
-                + ");");
     }
 }
