+++ src/main/java/com/xzavier0722/mc/plugin/slimefun4/storage/adapter/sqlite/SqliteAdapter.java
@@ -1,11 +1,8 @@
 package com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlite;
 
-import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.*;
-
 import city.norain.slimefun4.timings.entry.SQLEntry;
 import com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlCommonAdapter;
 import com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlUtils;
-import com.xzavier0722.mc.plugin.slimefun4.storage.common.DataScope;
 import com.xzavier0722.mc.plugin.slimefun4.storage.common.DataType;
 import com.xzavier0722.mc.plugin.slimefun4.storage.common.RecordKey;
 import com.xzavier0722.mc.plugin.slimefun4.storage.common.RecordSet;
@@ -91,249 +88,6 @@ public class SqliteAdapter extends SqlCommonAdapter<SqliteConfig> {
                 + SqlUtils.mapTable(key.getScope())
                 + SqlUtils.buildConditionStr(key.getConditions())
                 + ";");
-    }
-
-    private void createProfileTables() {
-        createProfileTable();
-        createResearchTable();
-        createBackpackTable();
-        createBackpackInvTable();
-    }
-
-    private void createBlockStorageTables() {
-        createBlockRecordTable();
-        createBlockDataTable();
-        createBlockInvTable();
-        createChunkDataTable();
-        createUniversalInventoryTable();
-        createUniversalRecordTable();
-    }
-
-    private void createProfileTable() {
-        var table = SqlUtils.mapTable(DataScope.PLAYER_PROFILE);
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + table
-                + "("
-                + FIELD_PLAYER_UUID
-                + " TEXT PRIMARY KEY NOT NULL, "
-                + FIELD_PLAYER_NAME
-                + " TEXT NOT NULL, "
-                + FIELD_BACKPACK_NUM
-                + " INTEGER DEFAULT 0"
-                + ");");
-
-        executeSql("CREATE INDEX IF NOT EXISTS index_player_name ON " + table + " (" + FIELD_PLAYER_NAME + ");");
-    }
-
-    private void createResearchTable() {
-        var table = SqlUtils.mapTable(DataScope.PLAYER_RESEARCH);
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + table
-                + "("
-                + FIELD_PLAYER_UUID
-                + " TEXT NOT NULL, "
-                + FIELD_RESEARCH_KEY
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "REFERENCES "
-                + SqlUtils.mapTable(DataScope.PLAYER_PROFILE)
-                + "("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE);");
-
-        executeSql("CREATE INDEX IF NOT EXISTS index_player_researches ON "
-                + table
-                + " ("
-                + FIELD_PLAYER_UUID
-                + ", "
-                + FIELD_RESEARCH_KEY
-                + ");");
-    }
-
-    private void createBackpackTable() {
-        var table = SqlUtils.mapTable(DataScope.BACKPACK_PROFILE);
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + table
-                + "("
-                + FIELD_BACKPACK_ID
-                + " TEXT PRIMARY KEY NOT NULL, "
-                + FIELD_PLAYER_UUID
-                + " TEXT NOT NULL, "
-                + FIELD_BACKPACK_NUM
-                + " INTEGER NOT NULL, "
-                + FIELD_BACKPACK_NAME
-                + " TEXT NULL, "
-                + FIELD_BACKPACK_SIZE
-                + " INTEGER NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "REFERENCES "
-                + SqlUtils.mapTable(DataScope.PLAYER_PROFILE)
-                + "("
-                + FIELD_PLAYER_UUID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE);");
-
-        executeSql("CREATE INDEX IF NOT EXISTS index_player_backpack ON "
-                + table
-                + " ("
-                + FIELD_PLAYER_UUID
-                + ", "
-                + FIELD_BACKPACK_NUM
-                + ");");
-    }
-
-    private void createBackpackInvTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + SqlUtils.mapTable(DataScope.BACKPACK_INVENTORY)
-                + "("
-                + FIELD_BACKPACK_ID
-                + " TEXT NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " INTEGER NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_BACKPACK_ID
-                + ") "
-                + "REFERENCES "
-                + SqlUtils.mapTable(DataScope.BACKPACK_PROFILE)
-                + "("
-                + FIELD_BACKPACK_ID
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_BACKPACK_ID
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createBlockRecordTable() {
-        var table = SqlUtils.mapTable(DataScope.BLOCK_RECORD);
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + table
-                + "("
-                + FIELD_LOCATION
-                + " TEXT PRIMARY KEY NOT NULL, "
-                + FIELD_CHUNK
-                + " TEXT NOT NULL, "
-                + FIELD_SLIMEFUN_ID
-                + " TEXT NOT NULL"
-                + ");");
-
-        executeSql("CREATE INDEX IF NOT EXISTS index_chunk ON " + table + "(" + FIELD_CHUNK + ");");
-    }
-
-    private void createBlockDataTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + SqlUtils.mapTable(DataScope.BLOCK_DATA)
-                + "("
-                + FIELD_LOCATION
-                + " TEXT NOT NULL, "
-                + FIELD_DATA_KEY
-                + " TEXT NOT NULL, "
-                + FIELD_DATA_VALUE
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_LOCATION
-                + ") "
-                + "REFERENCES "
-                + SqlUtils.mapTable(DataScope.BLOCK_RECORD)
-                + "("
-                + FIELD_LOCATION
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_LOCATION
-                + ", "
-                + FIELD_DATA_KEY
-                + ")"
-                + ");");
-    }
-
-    private void createChunkDataTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + SqlUtils.mapTable(DataScope.CHUNK_DATA)
-                + "("
-                + FIELD_CHUNK
-                + " TEXT NOT NULL, "
-                + FIELD_DATA_KEY
-                + " TEXT NOT NULL, "
-                + FIELD_DATA_VALUE
-                + " TEXT NOT NULL, "
-                + "PRIMARY KEY ("
-                + FIELD_CHUNK
-                + ", "
-                + FIELD_DATA_KEY
-                + ")"
-                + ");");
-    }
-
-    private void createBlockInvTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + SqlUtils.mapTable(DataScope.BLOCK_INVENTORY)
-                + "("
-                + FIELD_LOCATION
-                + " TEXT NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " INTEGER NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL, "
-                + "FOREIGN KEY ("
-                + FIELD_LOCATION
-                + ") "
-                + "REFERENCES "
-                + SqlUtils.mapTable(DataScope.BLOCK_RECORD)
-                + "("
-                + FIELD_LOCATION
-                + ") "
-                + "ON UPDATE CASCADE ON DELETE CASCADE, "
-                + "PRIMARY KEY ("
-                + FIELD_LOCATION
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createUniversalInventoryTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + SqlUtils.mapTable(DataScope.UNIVERSAL_INVENTORY)
-                + "("
-                + FIELD_UNIVERSAL_UUID
-                + " CHAR(64) NOT NULL, "
-                + FIELD_INVENTORY_SLOT
-                + " TINYINT UNSIGNED NOT NULL, "
-                + FIELD_INVENTORY_ITEM
-                + " TEXT NOT NULL,"
-                + "PRIMARY KEY ("
-                + FIELD_UNIVERSAL_UUID
-                + ", "
-                + FIELD_INVENTORY_SLOT
-                + ")"
-                + ");");
-    }
-
-    private void createUniversalRecordTable() {
-        executeSql("CREATE TABLE IF NOT EXISTS "
-                + SqlUtils.mapTable(DataScope.UNIVERSAL_RECORD)
-                + "("
-                + FIELD_UNIVERSAL_UUID
-                + " UUID NOT NULL, "
-                + FIELD_SLIMEFUN_ID
-                + " TEXT NOT NULL, "
-                + FIELD_LAST_PRESENT
-                + " TEXT NOT NULL, "
-                + "PRIMARY KEY ("
-                + FIELD_UNIVERSAL_UUID
-                + ")"
-                + ");");
     }
 
     public synchronized void executeSql(String sql) {
