+++ src/main/java/com/xzavier0722/mc/plugin/slimefun4/storage/controller/SlimefunUniversalData.java
@@ -2,6 +2,7 @@ package com.xzavier0722.mc.plugin.slimefun4.storage.controller;
 
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import java.util.UUID;
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import lombok.Getter;
@@ -20,6 +21,7 @@ public class SlimefunUniversalData extends ASlimefunDataContainer {
 
     @Setter
     @Getter
+    @Nonnull
     private volatile Location lastPresent;
 
     @Setter
